added TODO list and small fixes

+++ src/main.c
@@ -143,8 +143,8 @@ int main(int argc, char **argv)
       struct timeval curtime;
       gettimeofday(&curtime, NULL);
 
-      printf("TIME: %10d sec\n     %10d usec\n", curtime.tv_sec, curtime.tv_usec);
-
+      printf("TIME: %10d sec\n     %10d usec\n",
+				(unsigned int)curtime.tv_sec, (unsigned int)curtime.tv_usec);
 
       recv_msg(&pkt, &tval);
